CascadeFury/requirements.txt
python-binance==1.0.19
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
ccxt==4.1.13
ta==0.10.2
schedule==1.2.0
